containers/scraping/requirements.txt
webdriver_manager==4.0.0
undetected-chromedriver==3.5.2
bs4==0.0.1
sodapy==2.2.0
python-dotenv==1.0.0
PyMySQL==1.0.3
SQLAlchemy==2.0.16
pytz==2021.1
selenium==4.12.0
boto==2.49.0
python-dateutil==2.8.2
pandas==2.0.3
numpy==1.25.1